+++ tests/end2end/tests/url_param.rs
@@ -1,7 +1,7 @@
 use end2end_helpers::{element_text, input_by_id, mount};
 use leptos::prelude::*;
 use leptos_fluent::{leptos_fluent, url};
-use leptos_fluent_csr_minimal_example::{LanguageSelector, TRANSLATIONS};
+use leptos_fluent_csr_minimal_example::LanguageSelector;
 use wasm_bindgen_test::*;
 use web_sys_ec::{Ec, Wait};
 
@@ -13,7 +13,6 @@ const URL_PARAM: &str = "my-weird-url-param";
 fn I18n(children: Children) -> impl IntoView {
     leptos_fluent! {
         children: children(),
-        translations: [TRANSLATIONS],
         locales: "../../examples/csr-minimal/locales",
         initial_language_from_url_param: true,
         url_param: URL_PARAM,